requirements.txt
Django==4.2.*
djangorestframework==3.14.*
django-allauth==0.57.*
dj-database-url==2.1.*
python-decouple==3.8
psycopg2-binary==2.9.*
django-ratelimit==4.1.*
qrcode[pil]==7.4.*
Pillow==10.1.*
